+++ android/app/src/main/java/com/shahartal/myquietchannel/parasha/Item.kt
@@ -6,5 +6,6 @@ data class Item(
     val leyning: Leyning,
     val memo: String,
     val date: String,
+    // val title_orig: String,
     val link: String // https://www.sefaria.org/Zevachim.36a?lang=bi&utm_source=hebcal.com&utm_medium=api
 )