+++ android/app/src/main/java/com/shahartal/myquietchannel/parasha/Leyning.kt
@@ -1,5 +1,6 @@
 package com.shahartal.myquietchannel.parasha
 
 data class Leyning(
-    val haftarah: String
+    val haftarah: String,
+    var haftarah_sephardic: String
 )